Conflicts:
	Assets/NewAudioMixer.mixer.meta
	Assets/Scenes/UI_test.unity
	Assets/Scripts/THAW.meta
	Assets/Scripts/UI/BodyController.cs
	Assets/Sounds/Thumbs.db.meta

+++ Assets/Scripts/Body.cs
@@ -9,8 +9,8 @@ public class Body : MonoBehaviour {
 	public Vector3[] positions;
 	public Vector3[] velocities;
 
-	public float density;
-	public float volume;
+	public double density;
+	public double volume;
 	public float mass = 1;
 	public float radius;
 
@@ -124,21 +124,21 @@ public class Body : MonoBehaviour {
 	//////////////////////////////////
 	//     Interaction Handling
 
-	public void updateMass(float value){
-		mass = value;
+	public void updateMass(double value){
+		mass = (float)(value);
 	}
 
-	public void updateRadius(float value){
-		radius = value;
-		float dia = radius * 2;
+	public void updateRadius(double value){
+		radius = (float)(value);
+		float dia = (float)(radius * 2);
 		transform.localScale = new Vector3(dia,dia,dia);
 	}
 
-	public void updateVolume(float value){
+	public void updateVolume(double value){
 		volume = value;
 	}
 
-	public void updateDensity(float value){
+	public void updateDensity(double value){
 		density = value;
 	}
